pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = ["Rany Daioub <royrootdev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
python-telegram-bot = "21.6.*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
